refactor: 세션 페이지의 사이드바 컴포넌트를 분리
